# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SemIf decision service",
    "version": "0.1.0",
    "description": "Serves one conditional answer-slot distribution per request over a chat-completions-shaped route, whose caller must declare the decision schema. The returned probabilities are conditional option scores, not calibrated decision confidence; every response says so in `choices[].semif.probability_status`. A multi-field decision is not supported: one request scores one field, so N fields are N requests (or one `/v1/semif/batch`). Two statuses are router-level rather than per-path and are therefore not repeated under every path: `404 not_found` for a path this server does not answer, and `405 method_not_allowed` for a method it does not answer there. Every route except `GET /health` is gated by `--api-key` when set, and every non-preflight request answers `403 origin_not_allowed` when `--cors-origin` is an exact origin that the request's `Origin` does not match. A preflight `OPTIONS` is answered `204` before both checks, because a browser attaches neither the bearer token nor a CORS decision to it."
  },
  "servers": [{ "url": "/" }],
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness and the exact serving configuration",
        "description": "Never requires the API key, so a supervisor can probe liveness without a secret.",
        "responses": {
          "200": {
            "description": "The process is serving",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Health" }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "The served model, for OpenAI-client compatibility",
        "responses": {
          "200": {
            "description": "One model entry",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelList" }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "Score one decision",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompletionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The decision. `text/event-stream` when `stream` is true, with exactly one content chunk, one stop chunk, an optional usage chunk, then `data: [DONE]`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Completion" }
              },
              "text/event-stream": { "schema": { "type": "string" } }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/OriginRefused" },
          "404": { "$ref": "#/components/responses/ModelNotFound" },
          "413": { "$ref": "#/components/responses/TooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/BackendFailed" },
          "503": { "$ref": "#/components/responses/ClientGone" },
          "504": { "$ref": "#/components/responses/BackendTimeout" }
        }
      }
    },
    "/v1/semif/batch": {
      "post": {
        "summary": "Score N independent decisions in one round trip",
        "description": "An envelope for independent requests, not a multi-field readout: every item runs the same path a standalone request would, and each result carries its own status. The per-decision identity and provenance are per item.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["requests"],
                "properties": {
                  "requests": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 32,
                    "items": { "$ref": "#/components/schemas/CompletionRequest" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "One result per item, in order. The envelope is always 200; a failed item carries its own status and error body, so one bad item cannot hide the others.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/OriginRefused" },
          "413": { "$ref": "#/components/responses/TooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Process counters in Prometheus text format",
        "responses": {
          "200": {
            "description": "Counter and histogram exposition",
            "content": {
              "text/plain": { "schema": { "type": "string" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/OriginRefused" }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "This document",
        "responses": {
          "200": {
            "description": "The OpenAPI description of this service",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/OriginRefused" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required on every route except `/health` when the server was started with `--api-key`."
      }
    },
    "schemas": {
      "Message": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": { "type": "string", "description": "`system` and `developer` build the criterion; every other role is evidence." },
          "content": {
            "description": "A string, or an array of text and `image_url` parts. Any other part type is refused by name. An `image_url` part is refused unless the server was started with --allow-media (it switches to a degraded, server-tokenized readout); media in a `system`/`developer` message is always refused.",
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["type"],
                  "properties": {
                    "type": { "type": "string", "enum": ["text", "image_url"] },
                    "text": { "type": "string" },
                    "image_url": {
                      "type": "object",
                      "required": ["url"],
                      "properties": {
                        "url": {
                          "type": "string",
                          "description": "A `data:image/*;base64,...` URI (fingerprinted and size-capped) or, only with --allow-remote-media, an http(s) URL. The string is forwarded to the backend byte for byte."
                        }
                      }
                    }
                  }
                }
              }
            ]
          }
        }
      },
      "CompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "model": { "type": "string", "description": "Must equal the served model, or the request is 404." },
          "messages": {
            "type": "array",
            "minItems": 1,
            "maxItems": 512,
            "items": { "$ref": "#/components/schemas/Message" },
            "description": "The canonical `[system, user]` pair keeps the original prompt byte for byte (`evidence_renderer: system+user`); any other shape folds into a role-labelled transcript (`evidence_renderer: transcript`)."
          },
          "response_format": { "$ref": "#/components/schemas/ResponseFormat" },
          "schema": { "type": "object", "description": "Alternative to `response_format.json_schema.schema`." },
          "guided_json": { "type": "object", "description": "The vLLM/SGLang native spelling. Must be the object itself; a path or registered name is refused." },
          "logprobs": { "type": "boolean", "description": "Fills the standard `choices[].logprobs.content` with the per-slot logprobs this readout already computes. The token is the answer-slot letter." },
          "top_logprobs": { "type": "integer", "minimum": 0, "maximum": 20, "description": "How many ranked slots to include in `top_logprobs`." },
          "stream": { "type": "boolean", "description": "Answer as Server-Sent Events. The decision is not incremental: the stream carries exactly one content chunk." },
          "stream_options": {
            "type": "object",
            "properties": { "include_usage": { "type": "boolean" } },
            "description": "Requires `stream: true`; `include_usage` appends a usage-only chunk before `[DONE]`."
          },
          "n": { "type": "integer", "enum": [1], "description": "Only 1: this endpoint scores one distribution, it does not sample." },
          "max_tokens": { "type": "integer", "minimum": 1, "description": "Accepted and echoed in `ignored_parameters`; exactly one position is always returned." },
          "temperature": { "type": "number", "description": "Accepted and echoed in `ignored_parameters`; the readout is deterministic." }
        }
      },
      "ResponseFormat": {
        "type": "object",
        "required": ["type", "json_schema"],
        "properties": {
          "type": { "type": "string", "enum": ["json_schema"] },
          "json_schema": {
            "type": "object",
            "required": ["schema"],
            "properties": { "schema": { "$ref": "#/components/schemas/DecisionSchema" } }
          }
        }
      },
      "DecisionSchema": {
        "type": "object",
        "description": "Exactly one field, whose answer set is the option list. Slot order is the schema's enum order: the first value is slot A.",
        "required": ["type", "properties", "required", "additionalProperties"],
        "properties": {
          "type": { "type": "string", "enum": ["object"] },
          "properties": {
            "type": "object",
            "minProperties": 1,
            "maxProperties": 1,
            "description": "One field. Its `enum` or `type` defines the options."
          },
          "required": { "type": "array", "items": { "type": "string" } },
          "additionalProperties": { "type": "boolean", "enum": [false] }
        }
      },
      "Completion": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Deterministic in the prompt hash, so the same decision has the same id." },
          "object": { "type": "string", "enum": ["chat.completion"] },
          "created": { "type": "integer" },
          "model": { "type": "string" },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer" },
                "message": {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string", "enum": ["assistant"] },
                    "content": { "type": "string", "description": "The chosen option as JSON, e.g. `{\"verdict\":false}`." }
                  }
                },
                "finish_reason": { "type": "string", "enum": ["stop"] },
                "logprobs": { "$ref": "#/components/schemas/Logprobs" },
                "semif": { "$ref": "#/components/schemas/SemifBlock" }
              }
            }
          },
          "usage": { "$ref": "#/components/schemas/Usage" },
          "semif": { "$ref": "#/components/schemas/ProvenanceBlock" }
        }
      },
      "Logprobs": {
        "type": ["object", "null"],
        "description": "Null unless `logprobs: true`.",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "token": { "type": "string", "description": "The answer-slot letter, which is the token the engine scores." },
                "logprob": { "type": "number" },
                "bytes": { "type": "array", "items": { "type": "integer" } },
                "top_logprobs": { "type": "array", "items": { "type": "object" } }
              }
            }
          }
        }
      },
      "SemifBlock": {
        "type": "object",
        "description": "Everything needed to attribute this decision. `probabilities` and `option_logprobs` are parallel to `values`.",
        "properties": {
          "field": { "type": "string" },
          "values": { "type": "array", "description": "Schema order: index 0 is slot A." },
          "labels": { "type": "array", "items": { "type": "string" } },
          "choice": { "description": "The chosen value." },
          "choice_index": { "type": "integer" },
          "probabilities": { "type": "array", "items": { "type": "number" }, "description": "Conditional option scores; uncalibrated as decision confidence." },
          "option_logprobs": { "type": "array", "items": { "type": "number" } },
          "answer_token_ids": { "type": "array", "items": { "type": "integer" } },
          "option_ids": { "type": "array", "items": { "type": "string" } },
          "readout": { "type": "string", "description": "Which backend path produced the numbers, in words." },
          "fallback_used": { "type": "boolean", "description": "True when a slot missed the backend's top-N and /generative_scoring was used. That rate rises with load." },
          "endpoint": { "type": "integer", "description": "Index into the configured endpoint list." },
          "input_tokens": { "type": ["integer", "null"] },
          "input_ids_sha256": { "type": ["string", "null"], "description": "Text path only: sha256 of `json.dumps(input_ids)`. Null on the media path, where the prompt is server-tokenized." },
          "prompt_sha256": { "type": ["string", "null"], "description": "Text path only: sha256 of the rendered prompt. Null on the media path." },
          "request_sha256": { "type": ["string", "null"], "description": "Media path only: sha256 of the request body sent to the backend. Null on the text path." },
          "modality": { "type": "string", "enum": ["text", "text+image"], "description": "Media numbers must never be pooled with text numbers: the readout and serving_config differ." },
          "server_tokenized": { "type": "boolean", "description": "Always true: the token ids come from the backend on every path, so no local tokenizer can disagree with the engine that scores. It is therefore not what tells text from media; `modality` is. On the media path the backend's processor expands the image, so the answer-boundary proof does not apply there." },
          "media": { "type": "array", "items": { "type": "object" }, "description": "One provenance object per attached image: kind, mime, decoded bytes, and the sha256 of exactly those bytes (null for a remote URL, whose bytes this server never saw)." },
          "probability_status": { "type": "string" },
          "engine_seconds": { "type": "number" },
          "fallback_seconds": { "type": "number" },
          "total_seconds": { "type": "number" }
        }
      },
      "ProvenanceBlock": {
        "type": "object",
        "properties": {
          "backend": { "type": "string", "enum": ["vllm", "sglang"] },
          "serving_config": { "type": "string", "description": "The measured readout configuration. Text: `vllm-openai-logprob-token-ids-v1` (Tier A), `vllm-openai-completions-subset-softmax-v1` (the top-n path), `sglang-native-generate-token-ids-v1`. Media: `vllm-openai-chat-media-token-ids-v1` or `vllm-openai-chat-media-topn-v1`. Text and media numbers must never be pooled, and neither must two different text recipes." },
          "prompt_version": { "type": "string" },
          "prompt_template": { "type": "string" },
          "prompt_template_source": { "type": "string" },
          "revision": { "type": "string", "description": "The model revision, as given to `--revision`." },
          "prompt_contract": { "type": "string" },
          "evidence_renderer": { "type": "string", "enum": ["system+user", "transcript", "system+user+media", "transcript+media"], "description": "How `messages` were folded into the scored evidence. A given request must produce the same prompt within one value of this key. The `+media` values are a different prompt and must not be compared with the text values." },
          "ignored_parameters": { "type": "array", "items": { "type": "string" }, "description": "Parameters the caller sent that were accepted but had no effect. Never silent." },
          "client": { "type": "string" },
          "tokenize_endpoint": { "type": "string", "description": "The backend route this deployment tokenizes through: `POST /tokenize` (vLLM) or `POST /v1/tokenize` (SGLang). There is no local tokenizer, so this is what must answer for any decision to be scored." }
        }
      },
      "Usage": {
        "type": "object",
        "properties": {
          "prompt_tokens": {
            "type": ["integer", "null"],
            "description": "Null on a media request whose backend did not report `usage.prompt_tokens`: the client never saw the tokenized prompt, so it will not invent a count."
          },
          "completion_tokens": { "type": "integer", "enum": [1] },
          "total_tokens": { "type": ["integer", "null"] }
        }
      },
      "BatchResult": {
        "type": "object",
        "properties": {
          "object": { "type": "string", "enum": ["list"] },
          "count": { "type": "integer" },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "status": { "type": "integer", "description": "The status this item would have had on its own." },
                "body": { "description": "The item's completion object, or its error object." }
              }
            }
          }
        }
      },
      "Health": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["ok"] },
          "backend": { "type": "string" },
          "tier_a": { "type": "boolean" },
          "media_support": { "type": "string", "enum": ["none", "chat-top-n", "chat-exact-slot"], "description": "What the startup probe proved about scoring an image. `none` means an image request is refused; a refusal message may point a caller here." },
          "serving_config": { "type": "string" },
          "media_serving_config": { "type": ["string", "null"], "description": "The media readout recipe, null when this deployment does not serve media." },
          "prompt_version": { "type": "string" },
          "prompt_template": { "type": "string" },
          "prompt_template_source": { "type": "string" },
          "revision": { "type": "string" },
          "tokenizer_matches_served_root": { "type": ["boolean", "null"], "description": "Whether the served checkpoint's directory name matches the local `--model` directory. Null when the two could not be compared: no local checkpoint was named, or the backend reported no model root." },
          "client": { "type": "string" },
          "tokenize_endpoint": { "type": "string", "description": "The backend route this deployment tokenizes through: `POST /tokenize` (vLLM) or `POST /v1/tokenize` (SGLang). There is no local tokenizer, so this is what must answer for any decision to be scored." },
          "endpoints": { "type": "integer" },
          "endpoint_urls": { "type": "array", "items": { "type": "string" } },
          "served_model": { "type": "string" },
          "max_model_len": { "type": ["integer", "null"] },
          "media": {
            "type": "object",
            "description": "The media ceilings in force, so both limits are discoverable: the per-image one and the derived request-body one.",
            "properties": {
              "allow_media": { "type": "boolean" },
              "allow_remote": { "type": "boolean" },
              "max_images": { "type": "integer" },
              "max_media_bytes": { "type": "integer" }
            }
          },
          "max_body_bytes": { "type": "integer" }
        }
      },
      "ModelList": {
        "type": "object",
        "properties": {
          "object": { "type": "string", "enum": ["list"] },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "object": { "type": "string", "enum": ["model"] },
                "created": { "type": "integer" },
                "owned_by": { "type": "string" }
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["message", "type", "code"],
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string", "enum": ["invalid_request_error", "server_error"] },
              "code": { "type": "string" },
              "param": { "type": "string", "description": "The offending request field, when there is one. The difference between a caller that can fix its request and one that can only read prose." }
            }
          }
        }
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "The request cannot be answered faithfully. `error.param` names the field.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "`--api-key` is set and the bearer token is missing or wrong.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "OriginRefused": {
        "description": "`--cors-origin` is set to an exact origin and this browser origin is not it.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ModelNotFound": {
        "description": "`model` names something other than the served model.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "TooLarge": {
        "description": "The body exceeds the server's limit.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "BackendFailed": {
        "description": "The backend refused, broke, or answered unusably. Retrying is not obviously wrong, but the cause is not a spent budget.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "BackendTimeout": {
        "description": "The backend used the whole budget without answering, so the caller may reasonably retry.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "InternalError": {
        "description": "This server found itself inconsistent, e.g. a readout whose arity does not match the declared slots. The response body says what disagreed.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ClientGone": {
        "description": "The request was cancelled while it waited for a worker slot, so no backend readout was spent on it. A `/v1/semif/batch` item carries this status inside the envelope rather than as the response status.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}